[MOD/#15] 패키지 네임 수정
